removed kysely and related packages, moved to prisma client and adapter, updated related env, added signIn callback to nextAuth config to keep profile data in db in sync with spotify auth api, folder reorg
