+++ .env.template
@@ -1,12 +1,4 @@
 AUTH_SECRET=""
 AUTH_SPOTIFY_ID=""
 AUTH_SPOTIFY_SECRET=""
-
-# For Kysely auth adapter, used in /src/db/db.ts
-DATABASE_HOST=""
-DATABASE_NAME=""
-DATABASE_USER=""
-DATABASE_PASSWORD=""
-
-# Needed for prisma migrations, referenced in /prisma/schema.prisma
 DATABASE_URL=""
