+++ package.json
@@ -9,10 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@auth/kysely-adapter": "^1.1.0",
+    "@auth/prisma-adapter": "^2.2.0",
     "@fortawesome/fontawesome-svg-core": "^6.5.2",
     "@fortawesome/free-brands-svg-icons": "^6.5.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@prisma/client": "^5.15.0",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
@@ -26,13 +27,10 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "jotai": "^2.8.0",
-    "kysely": "^0.26.3",
     "lucide-react": "^0.373.0",
     "next": "^14.2.3",
     "next-auth": "^5.0.0-beta.18",
     "next-themes": "^0.3.0",
-    "pg": "^8.11.5",
-    "prisma-kysely": "^1.8.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-resizable-panels": "^2.0.19",
@@ -43,7 +41,6 @@
   "devDependencies": {
     "@types/eslint": "^8.56.10",
     "@types/node": "^18.16.1",
-    "@types/pg": "^8.11.6",
     "@types/react": "^18.2.75",
     "@types/react-dom": "^18.2.24",
     "@typescript-eslint/eslint-plugin": "^7.11.0",
@@ -52,7 +49,7 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "^14.2.3",
     "postcss": "^8.4.38",
-    "prisma": "^5.14.0",
+    "prisma": "^5.15.0",
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.4"
   }
