+++ prisma/schema.prisma
@@ -1,15 +1,32 @@
 // This is your Prisma schema file,
 // learn more about it in the docs: https://pris.ly/d/prisma-schema
 
-generator kysely {
-  provider    = "prisma-kysely"
-  output      = "../src/db"
-  readOnlyIds = true
-}
-
 datasource db {
   provider = "postgresql"
   url      = env("DATABASE_URL")
+}
+
+generator client {
+  provider    = "prisma-client-js"
+}
+
+model User {
+  id                String     @id @default(cuid())
+  spotifyId         String     @unique
+  name              String
+  email             String     @unique
+  emailVerified     DateTime?
+  images            Json[]
+  followers         Int
+  country           String
+  product           String
+  explicitFiltered  Boolean
+  explicitLocked    Boolean
+  accounts          Account[]
+  sessions          Session[]
+
+  createdAt         DateTime   @default(now())
+  updatedAt         DateTime   @updatedAt
 }
 
 model Account {
@@ -26,7 +43,7 @@ model Account {
   session_state      String?
 
   createdAt          DateTime  @default(now())
-  updatedAt          DateTime  @updatedAt @default(now())
+  updatedAt          DateTime  @updatedAt
 
   user               User      @relation(fields: [userId], references: [id], onDelete: Cascade)
 
@@ -37,34 +54,16 @@ model Session {
   sessionToken  String    @unique
   userId        String
   expires       DateTime
+  
+  createdAt     DateTime  @default(now())
+  updatedAt     DateTime  @updatedAt
+
   user          User      @relation(fields: [userId], references: [id], onDelete: Cascade)
 }
 
-model User {
-  id                String     @id @default(dbgenerated("gen_random_uuid()"))
-  spotifyId         String     @unique
-  name              String
-  email             String     @unique
-  emailVerified     DateTime?
-  /// @kyselyType({ url: string, height: number, width: number })
-  images            Json[]
-  followers         Int
-  country           String
-  /// @kyselyType("premium" | "free" | "open")
-  product           String
-  explicitFiltered  Boolean
-  explicitLocked    Boolean
-  accounts          Account[]
-  sessions          Session[]
-
-  createdAt         DateTime   @default(now())
-  updatedAt         DateTime   @updatedAt @default(now())
-}
-
-// Not used, not expected to be used, NextAuth wants it for KyselyAuth adapter type checking
 model VerificationToken {
   identifier  String
-  token       String    @unique
+  token       String
   expires     DateTime 
 
   @@unique([identifier, token])
