+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handlers } from "@/auth/auth";
+import { handlers } from "@/server/auth/auth";
 
 export const { GET, POST } = handlers;
