+++ src/server/db.ts
@@ -1,0 +1,9 @@
+import { PrismaClient } from "@prisma/client";
+import { env } from "@/env";
+
+export const db = new PrismaClient({
+  log:
+    env.NODE_ENV === "development" 
+      ? ["query", "error", "warn"] 
+      : ["error"],
+});
